Add new matsubara sum functions and update main programm
There is now a numerics matsubara sum formula in 'function2.py' and
a analytic formula in 'function3.py'. The analytic formula gives correct
values but stucks because of divide by zero errors.
phaseDiagramParallel.py has been updated such that it catches (mu, T)
values for which integration to k_IR doesn't work.
Still, the phase diagram looks not quite right although the shape is
somewhat similar to the Strodthoff paper.